add unit tests for moves in check

+++ src/bitboards.h
@@ -93,7 +93,7 @@ inline FLD PopLSB(U64& b)
 inline int CountBits(U64 b)
 {
 #if _WIN32 || _WIN64
-    return __popcnt64(b);
+    return static_cast<int>(__popcnt64(b));
 #else
     return __builtin_popcountl(b);
 #endif
